requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.1
python-dotenv==1.0.1
selenium==4.22.0
webdriver-manager==4.0.1
